pyproject.toml
[tool.poetry]
name = "pydualsense"
version = "0.7.2"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hidapi-usb = "^0.3.1"
pydantic = "^2.10.6"
ruff = "^0.9.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.0.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
poethepoet = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E402"]

[tool.ruff]
line-length = 120
exclude = ["**.pyi", "**/.venv/**"]
include = ["/**.py", "test/**.py"]
